backend/requirements.txt
Flask==2.3.0
pdfplumber==0.5.28
python-docx==0.8.11
spacy==3.4.3
scikit-learn==1.2.2
flask-cors==5.0.0